removed some redundancies
