+++ experiments/run_hrl.py
@@ -5,73 +5,9 @@ from time import strftime
 import sys
 
 from hbaselines.utils.misc import ensure_dir
-from hbaselines.utils.train import parse_options, get_hyperparameters
-from hbaselines.algorithms import RLAlgorithm
-
-EXAMPLE_USAGE = 'python run_hrl.py "HalfCheetah-v2" --meta_period 10'
-
-
-def run_exp(env,
-            policy,
-            hp,
-            dir_name,
-            evaluate,
-            seed,
-            eval_interval,
-            log_interval,
-            save_interval,
-            initial_exploration_steps,
-            ckpt_path):
-    """Run a single training procedure.
-
-    Parameters
-    ----------
-    env : str or gym.Env
-        the training/testing environment
-    policy : type [ hbaselines.base_policies.Policy ]
-        the policy class to use
-    hp : dict
-        additional algorithm hyper-parameters
-    dir_name : str
-        the location the results files are meant to be stored
-    evaluate : bool
-        whether to include an evaluation environment
-    seed : int
-        specified the random seed for numpy, tensorflow, and random
-    eval_interval : int
-        number of simulation steps in the training environment before an
-        evaluation is performed
-    log_interval : int
-        the number of training steps before logging training results
-    save_interval : int
-        number of simulation steps in the training environment before the model
-        is saved
-    initial_exploration_steps : int
-        number of timesteps that the policy is run before training to
-        initialize the replay buffer with samples
-    ckpt_path : str
-        path to a checkpoint file. The model is initialized with the weights
-        and biases within this checkpoint.
-    """
-    eval_env = env if evaluate else None
-
-    alg = RLAlgorithm(
-        policy=policy,
-        env=env,
-        eval_env=eval_env,
-        **hp
-    )
-
-    # perform training
-    alg.learn(
-        log_dir=dir_name,
-        log_interval=log_interval,
-        eval_interval=eval_interval,
-        save_interval=save_interval,
-        initial_exploration_steps=initial_exploration_steps,
-        seed=seed,
-        ckpt_path=ckpt_path,
-    )
+from hbaselines.utils.train import parse_options
+from hbaselines.utils.train import get_hyperparameters
+from hbaselines.utils.train import run_exp
 
 
 def main(args, base_dir):
@@ -134,7 +70,8 @@ if __name__ == '__main__':
         parse_options(
             description='Test the performance of goal-conditioned '
                         'hierarchical models on various environments.',
-            example_usage=EXAMPLE_USAGE,
+            example_usage='python run_hrl.py "HalfCheetah-v2" '
+                          '--meta_period 10',
             args=sys.argv[1:],
             hierarchical=True,
             multiagent=False,
